+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 2.7.8
 
 * Better automation of spell effects and mechanism for building in non-SRD spell effects if midi-qol is installed.
-* Flase Life will now roll correct hp as it scales.
+* False Life will now roll correct hp as it scales.
+* Spell activation conditions now added to the activation field.
 
 # 2.7.7